# defay1x9/api/openapi.yaml
openapi: 3.1.0
info:
  title: QuizBattle Gateway API
  version: 0.3.0
  description: HTTP and WebSocket contract for gateway, gameplay, auth, profile, bots and admin APIs.
servers:
  - url: http://localhost:8080
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: QB-SESSION
    csrfHeader:
      type: apiKey
      in: header
      name: X-CSRF-Token
  schemas:
...
    QuizTelegramErrorResponse:
      type: object
      required: [status, code, note, details]
      properties:
        status:
          type: string
          enum: [error]
        code:
          type: string
          example: validation_error
        note:
          type: string
          example: request validation failed
        details:
          type: object
          additionalProperties: true
...
paths:
...
  /api/v1/telegram/webhook/{botId}/{secret}:
    post:
      tags: [Telegram]
      summary: Unified Telegram webhook ingress (gateway-owned, no CSRF/cookie required)
      security: []
      parameters:
        - name: botId
          in: path
          required: true
          schema: { type: string }
        - name: secret
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        '200':
          description: Accepted
        '400':
          description: Invalid telegram update JSON
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QuizTelegramErrorResponse'
        '403':
          description: Invalid webhook secret (path or header)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QuizTelegramErrorResponse'
        '404':
          description: Unknown bot
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QuizTelegramErrorResponse'
...